+++ src/commands/registerContextMenuCommands.ts
@@ -1,7 +1,8 @@
 import * as vscode from 'vscode';
 import { commandPrefix } from '../extension';
+import { DartMultiplePubspecsWorkspaceType, DartNoPubspecWorkspaceType, DartSinglePubspecWorkspaceType, DartWorkspaceType, PubspecFile } from '../utils/analyzeWorkspaceType';
 import { createTerminal, runBuildRunner } from '../utils/terminalUtils';
-import { collectFiltersWithProgress, resolveUris } from '../utils/uri_utils';
+import { collectFilesToBuildWithProgress, resolveUris } from '../utils/uri_utils';
 
 export enum DartCommandType {
     Build = "build",
@@ -9,14 +10,14 @@ export enum DartCommandType {
 }
 
 export const contextMenuCommands = [
-    { id: 'buildThisFile', type: DartCommandType.Build, isPartFiles: false, title: "Build this file" },
-    { id: 'watchThisFile', type: DartCommandType.Watch, isPartFiles: false, title: "Watch this file" },
-    { id: 'buildPartFiles', type: DartCommandType.Build, isPartFiles: true, title: "Build part files" },
-    { id: 'watchPartFiles', type: DartCommandType.Watch, isPartFiles: true, title: "Watch part files" }
+    { id: 'buildThisFile', type: DartCommandType.Build, isPartFile: false, title: "Build this file" },
+    { id: 'watchThisFile', type: DartCommandType.Watch, isPartFile: false, title: "Watch this file" },
+    { id: 'buildPartFiles', type: DartCommandType.Build, isPartFile: true, title: "Build part files" },
+    { id: 'watchPartFiles', type: DartCommandType.Watch, isPartFile: true, title: "Watch part files" }
 ];
 
 export function registerFileCommands(context: vscode.ExtensionContext) {
-    contextMenuCommands.forEach(({ id, type, isPartFiles }) => {
+    contextMenuCommands.forEach(({ id, type, isPartFile: isPartFile }) => {
         context.subscriptions.push(vscode.commands.registerCommand(`${commandPrefix}.${id}`, async (file?: vscode.Uri, selectedFiles?: vscode.Uri[]) => {
             const uris = resolveUris(file, selectedFiles);
             if (uris.length === 0) {
@@ -24,27 +25,89 @@ export function registerFileCommands(context: vscode.ExtensionContext) {
                 return;
             }
 
-            const buildFilters = await collectFiltersWithProgress(uris, isPartFiles);
-            if (buildFilters.length > 0) {
-                runDartCommand(buildFilters, type);
+            const filesToBuild = await collectFilesToBuildWithProgress(uris, isPartFile);
+            if (filesToBuild.length > 0) {
+                runDartCommandFromContextMenu(context, filesToBuild, type);
             } else {
-                vscode.window.showWarningMessage("No part files found");
+                vscode.window.showWarningMessage("No dart files found");
             }
         }));
     });
 }
 
-function runDartCommand(files: string[], commandType: DartCommandType) {
-    const newTerminal = createTerminal(
-        files,
-        commandType,
-    );
+async function runDartCommandFromContextMenu(context: vscode.ExtensionContext, files: string[], commandType: DartCommandType) {
+    const pubspecFileDartFilesMap = await mapDartFilesToPubspecFiles(context, files);
 
-    runBuildRunner(
-        newTerminal,
-        files,
-        commandType,
-    );
+    if (pubspecFileDartFilesMap.size === 0) {
+        const filesStr = files.length === 1 ? 'file' : 'files';
+        vscode.window.showWarningMessage(`Selected ${filesStr} could not be built. Most likely there is no suitable pubspec.yaml found.`);
+        return;
+    }
 
-    newTerminal.show();
+    for (const [pubspec, dartFiles] of pubspecFileDartFilesMap) {
+        const newTerminal = createTerminal(
+            dartFiles,
+            commandType,
+            true,
+            pubspec
+        );
+
+        runBuildRunner(
+            newTerminal,
+            dartFiles,
+            commandType,
+            pubspec
+        );
+
+        newTerminal.show();
+    }
+}
+
+async function mapDartFilesToPubspecFiles(context: vscode.ExtensionContext, files: string[]): Promise<Map<PubspecFile, string[]>> {
+    const pubspecFileDartFilesMap = new Map<PubspecFile, string[]>();
+    const workspaceType = await DartWorkspaceType.getFromContext(context);
+    if (!workspaceType) {
+        return pubspecFileDartFilesMap;
+    }
+
+    const pubspecFiles: PubspecFile[] = [];
+    switch (workspaceType.constructor) {
+        case DartNoPubspecWorkspaceType:
+            return pubspecFileDartFilesMap;
+        case DartSinglePubspecWorkspaceType:
+            pubspecFiles.push((workspaceType as DartSinglePubspecWorkspaceType).pubspec);
+            break;
+        case DartMultiplePubspecsWorkspaceType:
+            pubspecFiles.push(...(workspaceType as DartMultiplePubspecsWorkspaceType).pubspecs);
+            break;
+    }
+
+
+    for (const file of files) {
+        var parentPubspec: PubspecFile | undefined;
+        var libIndex = file.lastIndexOf('lib/');
+        if (libIndex === -1) {
+            continue;
+        }
+
+        while (libIndex !== -1) {
+            const potentialParentPubspecPath = '/' + file.substring(0, libIndex) + 'pubspec.yaml';
+            parentPubspec = pubspecFiles.find(pubspec => pubspec.workspaceUri.fsPath === potentialParentPubspecPath);
+            if (parentPubspec) {
+                break;
+            }
+
+            libIndex = file.lastIndexOf('lib/', libIndex - 1);
+        }
+
+        if (!parentPubspec) {
+            continue;
+        }
+
+        const dartFiles = pubspecFileDartFilesMap.get(parentPubspec) || [];
+        dartFiles.push(file);
+        pubspecFileDartFilesMap.set(parentPubspec, dartFiles);
+    }
+
+    return pubspecFileDartFilesMap;
 }